+++ src/Symfony/Bundle/TwigBundle/Resources/config/twig.xml
@@ -52,7 +52,7 @@
             <tag name="twig.extension" />
         </service>
 
-        <service id="twig.extension.form" class="Symfony\Bridge\Twig\Extension\FormExtension" public="false">
+        <service id="twig.extension.form" class="Symfony\Bridge\Twig\Form\FormExtension" public="false">
             <tag name="twig.extension" />
             <argument>%twig.form.resources%</argument>
         </service>
